split styling into seperate files and additional adjustments to navbuttons
