+++ src/components/contact/Contact.css
@@ -1,0 +1,21 @@
+.contactContainer {
+    border:5px sold black;
+    height:100vh;
+    width:80vw;
+    background-color: white;
+    position:relative;
+    top:11vh;
+    margin:auto;
+    overflow:scroll;
+}
+
+.contactInfo {
+    height: 15vh;
+    width:75vw;
+    /* border:5px solid black; */
+    background-color: black;
+    color:white;
+    margin:auto;
+
+}
+
