+++ src/components/contact/Contact.js
@@ -1,0 +1,44 @@
+import React from 'react'
+import './Contact.css'
+
+const Contact = () => {
+  return (
+    <div className = 'contactContainer'>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+        <div className = 'contactInfo'>
+            <p className='contactContent'>Hello</p>
+        </div>
+
+    </div>
+  )
+}
+
+export default Contact
