+++ src/components/navbar/Navbar.css
@@ -1,0 +1,34 @@
+
+.header {
+    height:10vh;
+    width:100vw;
+    background-color: black;
+    /* margin:auto; */
+    display:flex;
+    /* list-style: none; */
+    /* justify-content: space-between; */
+    align-items: center;
+    position:fixed;
+  
+  }
+  
+  .navbar {
+    display:flex;
+    margin-left:auto;
+  }
+  
+  
+  
+  .navbutton {
+    width:6vw;
+    height:1.5vw;
+    background-color: white;
+    justify-content: center;
+    display:flex;
+    align-items:center;
+    position:relative;
+    right:1vw;
+    margin: .3em;    
+    border: 5px solid red;
+  }
+
