+++ src/components/navbar/Navbar.js
@@ -1,7 +1,9 @@
-import React from 'react'
+import React, {useState} from 'react'
 import { Link } from 'react-router-dom'
+import './Navbar.css'
 
 const Navbar = () => {
+  
   return (
     <div className='header'>
       <>
