+++ src/components/resume/Resume.css
@@ -1,0 +1,18 @@
+.resumeContainer {
+    position:relative;
+    top:11vh;
+    border: 5px solid white;
+    height:80vh;
+    width:95vw;
+    margin:auto;
+    background-color: blue;
+  }
+  
+  .resumeInfo {
+    border: 5px solid black;
+    height:10vw;
+    width:75vw;
+    margin:auto;
+    background-color: white;
+    margin-top:1vh;
+  }
