+++ src/components/resume/Resume.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Resume.css';
 
 const Resume = () => {
   return (
